bugfix shift overflow

+++ sa-builder/src/lib.rs
@@ -61,7 +61,7 @@ pub fn build_ssa(
         SAConstructionAlgorithm::LibSais => {
             let mut packed_text = bitpack_text(text, sparseness_factor);
             
-            libsais64_rs::sais64_long(&mut packed_text, 1 << (BITS_PER_CHAR * sparseness_factor as usize), sparseness_factor)
+            libsais64_rs::sais64_long(&mut packed_text, (1 as i64) << (BITS_PER_CHAR * sparseness_factor as usize), sparseness_factor)
         },
         SAConstructionAlgorithm::LibDivSufSort => libdivsufsort_rs::divsufsort64(text)
     }
